+++ packages/react/src/atomic/IconButton/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Icons are used to visually communicate core parts of the product and available actions. They can act as wayfinding tools to help merchants more easily understand where they are in the product, and common interaction patterns that are available.
 
+## 2025-08-21 `4.2.0`
+
+#### 🎉 New features
+
+- Added optional `color` prop to `IconButton` (defaults to `neutral-textHigh`) to allow customizing the inner `Icon` color. ([#334](https://github.com/TiendaNube/nimbus-design-system/pull/334) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-03-18 `4.1.2`
 
 ### 💡 Others